feat: create volume controls component
LP-7
