+++ src/features/Controls/components/VolumeControls/VolumeControls.module.scss
@@ -1,0 +1,9 @@
+@use '../../../../styles/palette' as colors;
+
+.volume {
+  &__container {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+  }
+}
